Move WrappedResult to dotty.util
We can't have WrappedResult in tools itself, since that gives a CyclicReference
error in some situations. Specifically, if tools.WrappedResult is compiled first,
we complete its paackage,
which means we complete tools/package.scala
which means we enter its members
which means we process the WrappedResult export
which means we complete WrappedResult
which gives a CyclicReference

An alternative fix would be to avoid the export and use explicit forwarders instead.
