+++ compiler/src/dotty/tools/dotc/core/Decorators.scala
@@ -11,6 +11,7 @@ import dotty.tools.dotc.transform.MegaPhase
 import ast.tpd._
 import scala.language.implicitConversions
 import printing.Formatting._
+import dotty.util.WrappedResult
 
 /** This object provides useful implicit decorators for types defined elsewhere */
 object Decorators {
